requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.11.0
numpy==1.18.4
psycopg2-binary==2.8.5
pytz==2020.1
scapy==2.4.3
scapy-http==1.8.2
sqlparse==0.3.1
